Add files via upload
Dinger Payment Gate Way(Test Case) for odoo version 18.
Requirement : To gather the response and sending format (fields) for the api. In my way: i am sending the payload with the user selected payment method and the form of sensitive input (phone, bank card number of customer) will show from the side of dinger api.
- Now ,Some code are commented to not work due to the response time or by not known of some api information.
-In the data of payment method.xml can add new payment method. 

Important: Payment provider is make with strongly to work with hook in the init file . when newly activate,upgrade or unstall , the module will not be accept. 
-To make it possible, the required payment method and the field of the payment or payment_transaction should be add firstly before the activate. Based on your modification, the computation or state change can make later.